Add data from 13 subjects.
These are the first subjects from the Spring 2014 PSYC 4802 class.  It
includes 13 of the 14 students running themselves.

The data files include both the demo run from the 2nd day of class, and a
testing run that can be treated like actual data.  To tell the difference,
look at the software version.  We used 0.98 on the 2nd day, and 1.0 for the
useable data.